fasta.py: fix doctest for hash randomization
Python 2.7 adds an option to randomize hash seeds, and this is enabled by
default in 3.3. The natural ordering of a dict now changes every time the
interpreter is started, so don't depend on this in a doctest.

Explicitly sort the items of a dict in order to compare against expected
values.