test_all.check_array_copy: fix NumPy type in Python 3
With the MemoryRecord subclass of FastaRecord, the sequence data is stored
internally as a str ('unicode' in Python 2). Making a NumPy array from this
data type defaults to encoding this str as UTF-32, which won't correctly
decode as ASCII or UTF-8.

Fix check_array_copy by explicitly telling np.array to use one byte per
character so that we can decode this with the system default
(ASCII-compatible) encoding.