Fix Fasta doctests for Python 2 and 3
The pyfasta code uses text strings (str in Python 3, unicode in Python 2)
internally in most places, and it's difficult to directly check the __repr__
of these objects against known values in a way that works in both versions.
unicode.__repr__ has a 'u' prefix in Python 2, but this is not present in
str.__repr__ in Python 3.

To get around this, explicitly call print() on everything that we're
checking.